Banking Project
Customer experience is an integral part of a bank’s operations. That’s why banks 
focus a lot on improving customer experience by removing hassles and 
enhancing the facilities they provide. Opening a new account in a bank usually 
requires a person to visit the bank, fill out a form, and submit the necessary 
papers. All of these tasks take up a lot of time and dampen the overall customer 
experience. Moreover, many people have to take time out of their schedules to 
go to a bank.